Allow for fate offsets when unique fates in use
Prior to this, fates were always numbered starting with 1 for protos
and starting with 0 for alts.  We are moving to a model in which all
fate numbers are unique across the process.  Every NFA now knows its
own range of fates, in the range of fbeg ..^ fend.  Most of this is
handled in nqp, but MoarVM has to know about fate offsets for dealing
with the results of alts.  Conveniently, for both protos and alts,
the 0th fate name entry was already unused and set to 0, so we encode
the new fbeg and fend into the 0th fate name.  (For alts, these fate
names were set to 0,1...*, and the array was only used by MoarVM to
determine how many different fates might be returned.  But importantly,
we have access to the array, so we can get easy access to fate name 0
without changing how the nqp::nfarunalt opcode takes arguments.)