Investigate conversion to two-operand form
x86 uses two-operand form, while the expression JIT assumes
three-operand form. This might cause problems if the second operand is
equal to the output operand and the first operand cannot be overwritten
and instruction is a direct non-commutative binary operator.

I'd rather have the register allocator resolve this somehow, but that's
rather complex as the full set of registers for all definitions of the
live range are not necessarily defined.