Implement configurable collation_mode for MVM_unicode_string_compare
This change causes the collation_mode argument to affect how the strings are
compared. Quaternary level will ignore differences in codepoints and only
rely on the collation values.

This also fixes a problem in interp.c where it did not pass the three integer arguments,
and also fixes the arguments being incorrectly set to MVMint32 instead of MVMint64