On Windows with MinGW toolchain, add -municode as late as possible
Without `-municode`, linking fails. However, if `-municode` is specified
before probing (either in the environment or in Configure.pl), the
probes fail because all of them use `main` rather than `wmain`.

Editing the probes would have disturbed `gcc` feature detection on other
platforms. My solution was to let all the probes run, and add
`-municode` to linker flags just before generating the `Makefile` when
using the MinGW toolchain.

I am not certain this is the right thing to do in a cross-compilaton
context.