fix GC problem in string repeat op
it set up the number of strands to be 1, but it didn't fill
the strands array with valid data. after that, it could get
into GC from collapse_strands, where it would then try to
follow pointers from the strands array, thus panicking.

samcv++ for the code that triggered this, dogbert17++ and jnthn++
for hunting and fixing respectively