Refactor spill_register
Because it is also necessary to be able to spill a specific
register (rather than just any), I need to reuse the logic for spilling
a live range, which has been extracted. The logic to find a spillable
position and to select the 'optimal' live range for spilling has also
been split out because they can develop independently.