Find 'special' tiles and stub handling them
Some of the infrastructure (notably, a map of register -> live range) to
actually compile CALL and ARGLIST is not yet available. But now we can
at least find these tiles.

That said, it might be nice to roll ARGLIST and CALL into a single tile
so we can ensure that loading the argument to CALL (sometimes a dynamic
pointer) will not conflict with call arguments set up by ARGLIST.