Investigate large constants
Large (64 bit) constants currently force a excessive expression node
size, and they typically need distinct handling from 'regular'
constants, esp. pointers, which we would like to have in a table for
linking, and for binary operators, in which directly using a 64 bit
constant is not directly supported on x86 while a 32 bit constant is. So
it would be interesting to split the large constants from the small
ones.