Break reference cycles in ARGLIST compilation
It is possible for cycles to occur in moving values towards their
call argument positions. In this case, we need to clear it in
reverse order of dependencies, which is what we achieve by a stack.

Still todo: correctly spilling registers that 'survive' the call,
correctly accounting for registers that are used by the call arg.