Implement ARGLIST/CALL conflict resolution
CALL may potentially use register values, which may conflict with the
arguments used by ARGLIST, we have to move them out of the way prior to
loading the ARGLIST values. Use a bitmap and (instrinic) find-first-set
operations to efficiently find the conflicting register and a free spot.
Will need to take care of potential CALL arg spilling; any values need
to be inserted prior to the ARGLIST, which is a pretty good argument for
merging CALL and ARGLIST.