Ignore SIGPIPE by default.
We already handle (or should handle) various errors. This addresses
issue #547, where a write to a process that does not exist would not
break the Promise of the write, but instead cause the program to exit
due to the SIGPIPE it received. This does not block users from adding
their own SIGPIPE handler.