Correctly detect+handle overflow in mp_get_int64
Handles the fact that for 64-bit 2's complement numbers the maximum
positive value is 2**63 - 1, but the maximum negative value is 2**63.

Uses the mp_get_long_long now available with the newer version of
libtommath. Also, instead of counting the bits every time, only do it
during the too-large-value error path.