Fix case insensitive string compare bug when synthetics in haystack
If we have a synthetic in the haystack, we have to do more work and
use the MVM_nfg_get_case_change function which will properly work with
synthetics.

This fixes a bug that would cause error "should eventually be unreachable"
on doing case insensitive regex or other case insensitive ops.