chr: For cp < 0x300 short circuit a unicode property test
If the codepoint decomposes we may need to normalize it.
The first cp that decomposes is U+0340, To be on the safe side, for
now we go with the first significant character which at the time of
writing is COMBINING GRAVE ACCENT U+300

This saves us a property check on the Decomposition_Type property.