Split value-nodes from void nodes
IF, DO and CALL would return a value only if their child nodes were
returning a value. (CALL had an parameter signifiying the 'return
type'). But because the tiler does not take such parameters into
account, this leads to confusing results. By splitting the
value-yielding from void uses of the nodes, each node now corresponds to
exactly one type, and no conflicts can occur. Although this increases
the number of nodes, it reduces the number of mechanisms, which is
pretty interesting.