Document automatic unconditionalizing
A value first computed in a conditional branch and used in a different
conditional branch or indeed in any basic block that is not followed
solely by the conditional is at risk of never being computed. We have
the 'let rooting hack' to ensure that the computation is ordered
before the conditional block (and hence available anywhere), but
that's hacky, error-prone and would not work for dynamically generated
tree fragments. So I've documented a way to order node computations
automatically.