Fix mp_get_int64 and mp_get_uint64 for realz
mp_get_long_long does in fact overflow/wrap around when the mp_int is
too large. So bring back checking the number of bits in the mp_int
before calling it. Also, remove the redundant check for whether the
mp_int is 0, since the exact same check is done in mp_get_long_long.