Propagate block numbers upwards
In a number of cases, the label and branch structure of a child node
is the same as the containing conditional structure, e.g. an ALL node
'flows' into its consequent block because it works by branching to the
alternative on any failed test. No basic block is started for them.
However, the block numbers are also required for correct assignment of
successors, and without uppropagation we might need to dig arbitrarily
deep to find them.