Make nursery cleanup non-concurrent, for now.
We can't both have thread-local free lists and have the possibility
of a work-stolen thread resuming while another thread that stole its
GC work frees its nursery, since they can race on the free list. It
is possible in the future to only lose the potential concurrency on
threads that were stolen, and let non-stolen ones go their own way as
before this patch. But for now:

* This has no negative impact for single-threaded applications
* The impact of the better FSA massively out-weighs the cost of this
  change for multi-threaded applications

We can likely do better here *and* have the better FSA, but that will
need a good bit more thought, care, and testing, while this change is
the "obviously safe" one that means we can merge the improvements to
the FSA with confidence.