Extend inner basic blocks with unconditional jump
In some cases, the (last) block associated with an 'inner' node
(e.g. nested ANY within WHEN or IFV) has the same successors as the
'outer' node; for example if there is an unconditional branch that
follows a conditional branch. Rather than introduce a new basic block
just for the the unconditional branch, this now includes the
unconditional branch in the last basic block.