Fix bug in index ignorecase and equal at ignorecase
This bug occurred when the entire length of the haystack was searched
and all of the graphemes matched the needle.

If the needle exceeded the length of the haystack past that point,
it would erroneously think there was a match there, as it only checked
that it matched the whole length of the haystack.

Would cause 'fgh' to be found in: 'abcdefg'. This only occurred at
the very end of the haystack.

The internal string_equal_at_ignore_case_INTERNAL_loop now returns
-1 if there was no match and 0 or more if there was a match at that index.

This return value provides new information which is 0 if there was a match
and some positive integer when the haystack was expanded when casefolding it.

This information had been planned to be exposed in some way at a future date,
as if we are searching for 'st' inside a string 'ﬆabc', index ignorecase
will indicate that it is located at index 0, and in Perl 6 Rakudo it will
return 'ﬆa' when it should instead have returned 'ﬆ'.

For now this additional information is only internal and the return values
of the indexic_s and equatic_s ops have not changed.