Fix off-by-one in block-generation
Range of basic block is [start, stop) so that we can
for (i = start; i < stop; i++) 'standard' C looping rather than
'i <= stop', which is really just as easy but more confusing.
Anyway, +1-ing start was wrong. Add some debug names in WHEN,
could do the same in other construct but not necessary yet.