Rename h_vars to H_vars to make haystack vars distinguishable from n_vars
We use n_graphs and h_graphs or n_cps etc. in many places.
h and n are not super easy to distinguish. Go with convention of using
a capital H for any uses of haystack and lowercase n for needle.

This makes it much easier to visually scan code and know which object
is which and the likely relative sizes each may be.