Find live range holes
Per Wimmer (2010), find 'holes' in live ranges (i.e. ranges within the
live range in which the value is not defined). Strategy is to iterate
backwards over all basic blocks and instructions. If a value is used,
that means it must be live in the code preceeding it (otherwise the
program is wrong). If a value is defined, that means it must be
undefined in code preceeding it, because all code is in SSA form. If a
value is used in one successor of the basic block but not in the
other, it must be live at the end of this basic block, because
otherwise it cannot be live in the successor... etc.

The goal is to save the register allocator from trying to restore a
value which isn't (yet) defined, potentially over the very definition
of that value.

I use a bitmap for managing the sets of live ranges, which is
reasonably memory-efficient in this case because live range indexes
are sequential integers, and pointers are large. However, I noticed
that global values also receive live ranges in this case, and that may
be wastefuls as these values never need to be allocated and cannot
have holes in them.