Actually use the bitmap sets correctly
The tile refs array is set up kind of weirdly; it's zero-biased and
includes only refs to nodes that it uses (excludes itself), which is
reasonable by itself, but it is at odds with the register requirements
bitmap, which does include the output register. So that may be target
for a cleanup.