Better version of patch to fix right shifting negative smallints
The previous patch did not check what sign was there and didn't fully
handle big/smallint containing P6int's. This one adds a new
BIGINT_IS_NEGATIVE() function to check if it is negative.

This function may be very useful other places we need to check if a
p6int is negative.