Remove empty live ranges from worklist
As a result of the merger of multiple live ranges, one live range ends
up as empty, which are defined as having a first reference of
MAX_INT32. This violated the initial heap structure of the
worklist which I had assumed was held by construction. A compaction
step has been added to remove empty live ranges and restore the
heap order.