Wrap invokish and throwish with a guard tile
Because this means that control flow can be interrupted and resumed
right after the compiled node, the expression assembler flushes all
active values after the guard, forcing them to be reloaded after.

The pre-throwish pseudotile (which is responsible for loading the JIT
entry label) requires at least two registers, one to hold the frame
and one to hold the entry label, so since we have just a single spare
register (rax), we must allocate one. This is kind of hackish and
non-portable, and it'd be nicer if we could make the tile template
compiler handle this.

This makes decont work correctly through CORE.setting compilation.