Invoke prepare_arglist_and_call for CALL tile
It could potentially happen that the register used by CALL would be
allocated after the ARGLIST was processed, e.g. when the associated
value is spilled and loaded by a synthetic tile. This could then
overwrite the register picked by prepare_arglist_and_call for
resolving the conflict between the ARGLIST and CALL registers.
Now that prepare_arglist_and_call is executed for the CALL
tile this can no longer happen.