Fix order of shift-in for CALL arg bitmap
The OR of the shifted-in value would be applied to the negated bitmask
of the shift-out value which would be AND-ed with the original
bitmap. The effect is that the shift-out bit would be unset, but the
shift-in bit would not be. This would cause that same register to be
reused as the 'spare' register for the chain-breaking cycle, while it
was really already in use.