Eliminate the use of the decode stream.
In doing so, introduce a simple 1-packet buffer in its place. In the
best case, programs will hit the no-memcpy path and the buffer we get
from the network will be handed straight on back to userspace after
being wrapped in an MVMArray. The buffer helps in cases where folks
do stuff like .read(1) on the socket to pull a byte off at a time.