Fix cycle break ordering
In ordering registers for function calls, it may occur that there is a
cycle in the transfer graph. These cycles must be broken somehow. I had
used a stack, but that would cause the cycle to be broken 'in reverse',
which would only be a problem if there were more than two registers
involved and therefore went unnoticed.

The simpler solution - walking the chain of transfers 'backwards' until
we reach the starting point - is actually correct.