Uses preceed definitions
Suppose a tile uses and defines a value at the same time, e.g.

$a = $a + $b

Then the definition of $a would open a hole, only for the use to close
it immediately. If it were the other way arround, the JIT would presume
a hole (backwards from $a) where there wasn't any. So this patch fixes
that.