Enable JIT breakpoint insertion
In order to stop a program at exactly the right location within a
JIT-compiled program, it is now possible to set an environment
variable MVM_JIT_BREAKPOINTS with the format frame_nr/block_nr, and
the JIT will kindly insert a breakpoint. This can then be used under a
debugger to step through the code.