[unicode-grant] Add improved support for GCB=Prepend
Previously we would only have the right functionality with
Grapheme_Cluster_Break=Prepend if the codepoint following the
Prepend character had NFG_QC=0.

We now save state in the MVMNormalizer so we know to not take the short
normalization path when we follow a Prepend character.

The only case this doesn't hold is when a control character follows a
Prepend character.