Remove point spilling for calls
Point spilling was supposed to be an optimization for rare function
call paths, but it interacted really poorly with full-spilling and
judging from the generated code was just as often a pessimization.
So I've decided to remove it, simplifying the ARGLIST preparing code,
but also unocvering a bug in handling spilled values for ARGLIST,
which I've also fixed.
I intend to optimize spilling separately, making sure a value is
not redundantly 'redefined' if the earlier definition is still
in the same basic block.