Clean up dead variables in PHI nodes.
Since those paths will never be taken, their facts are not relevant.
This means the merged facts can be more accurate. This allows for far
better handling of the case where an optional parameter was passed;
we only care about the facts on the variable, not those on the (dead)
default value.