Correct deopt check when removing named arg marks.
We only care about whether there is a deopt point during argument
processing. If there isn't, we can safely strip out the tracking.
It's only when there is that we must leave the tracking in place,
which in turn prevents inlining. Previously we tracked it there
was a deopt anywhere in the routine, rather than just up to the
end of argument processing.