Add a UTF-8 decoding fast-path.
In many cases, we are reading things that are already in normal form
thanks to being in the ASCII/Latin-1 range, which we can very cheaply
check. Add a fast path to the UTF-8 decoder for this case. It doesn't
actually skip anything UTF-8 related, but rather can take a shortcut
on the full normalizer. On the benchmark reading a million lines using
the UTF-8 encoding but with everything falling under the fast path, we
save around 3216 million CPU instructions.

When reading line by line, if the line separators are controls, this
mechanism ends up just naturally resetting itself back to trying the
fast path. This means that reading a file line by line, where just the
odd line has things needing full normalization, we'll still be able to
take the faster path on all of the other lines.