Make a smarter guess at decode result buffer size.
Previously we took the bytes available to decode, but when reading
lines there may be a thousand lines worth in the buffer. This avoids
allocating huge chunks of memory only to throw them away again soon
afterwards. It also is a prerequisite for just using those buffers in
the result string rather than doing copying.