Don't copy when we can steal decoder output.
When reading lines or chunks of chars, the decoder will often produce
a single buffer with the required content. In this case, just take it
and use it as the string's body, instead of copying it to another
buffer. This saves around 400 million instructions on the Perl 6 read
a million lines benchmark.