[unicode-grant] handle additional possible cases in collation
This change should handle cases where the a collation value
and b collation values match, and make sure we properly continue
through the string.

We also make sure to handle matching strings except for length.
If we have reversed quaternary enabled, we want to return
the 1 if string a is longer than string b.