[unicode-grant] store collation levels in an array not a single value
This fixes a bug where reversing secondary and tertiary levels could
cause the primary level to be overridden as a result. The Unicode
Collation Algorithm makes it very clear that this is not acceptable
functionality. The bug could cause ‘1’ to be sorted below
the ‘a’ due to their primary collation values being similar;
i.e. closer than the value of the secondary and tertiary levels.

This change puts the collation levels into an array instead of
a single value so we can easily iterate through them in order to
ensure higher levels do not override lower ones.

The type that stores the collation value is now an unsigned integer,
so that in the event of an overflow it cannot become negative.

Also remove an collation value check which is no longer needed.