Factor comparison out of live range heap
I want to maintain two different orders for the live range heap,
i.e. on the first_ref ascending order, and on last_ref ascending
order, for the values that are supposed to become active and for the
values that have been spilled respectively. So I've factored the
comparison function out of the live range heap.

A 'generic' binary heap in C would either require dynamic
allocation (to maintain space for the temporary when swapping) or be
written with macros. I don't find either solution very appealing.