Use jit code local types array for roots
Spilled values are stored, by default, in the frame work array.
If they had been on the stack, entering the interpreter would
destroy them.

This moves those spilled values into the work array proper (they used
to be after the args array) and ensures that the GC will properly
trace and update them. This then ensures that the JIT compiled code
will always have the proper values regardless of whether we call
something that allocates memory.