empty out bbs that were killed
they were still being kept in the tree made by "children" vertices,
and so other functions, most notably second_pass, would iterate over
them again, seeing already-updated usage counts for actually-dead
instructions.

In particular, the way prof_allocated was being handled caused
trouble here when it was in a killed BB and set the usage of something
to 1, then the dead BB was visited again and the usage count interpreted
as "only the prof_allocated instruction was using this" and therefor
thrown out, causing low-level nulls in registers to reach nasty spots
and causing segfaults.