Start giving frames a spesh log correlation ID.
Only when they are warm enough, and up to the point that they have
logged enough. Note that the invocations and spesh limit fields of
MVMStaticFrame will go away after the refactor, leaving it the same
size as they started. In MVMFrame, spesh_log_idx and osr_count will
be able to be removed, and we might back this new field 16 bits if
that will win better packing.