Update dynasm to get OPTREX fix
Some instructions only optionally need a REX byte. In this case, the
size of the generated code varies with the selected register, which
is known at link time. Previously we fixed that by increasing the size
if we found a REX byte was necessary. However, this interacts badly
with label offset calculation - specifically, we would allocate a single
byte for a relative label, which would then overflow if the code grew
due to addition of REX bytes. So instead, we now always assume that the
space is required, and decrease the offset if it turns out that it isn't.