Handle spesh annotations in the expr JIT
Previously we tried to let the legaacy JIT handle these, but that
doesn't work very well, as we might handle the same annotations twice,
and we didn't handle the after-instruction annotations at all. Also,
it lead to the 'breakup' of the expression tree.

With this patch the expr jit handles virtually all annotations, with
the exception of DEOPT_INLINE and DEOPT_ONE, because these are
attached to guards anyway, and we don't handle those yet.