Make illegal synthetic an oops, not a throw.
This partially undoes f621f21191c5c3a3035d3121c042cba4348c18cc, but
leaves in place the stack trace. The reasoning is that an invalid
synthetic is almost certainly going to be an indication of a serious
internal error (potentially corruption); only MoarVM constructs
synthetic codepoints, so any time an invalid one shows up it has to
be a VM mistake.