Work Hangul grapheme seg into GCB switches
This simplifies the code in should_break() and we are able to
use the same switches we use for all the other properties.

This mostly has the same functionality except it is more strictly in
line with Unicode's spec. by going off property and not assuming
everything returning true from maybe_hangul() is Hangul.