invokewithcapture must keep the used capture alive
The frame's argument processing context will come to reference the
effective callsite. If it were to be collected before the frame was
executed, then the frame would get a dangling callsite pointer. This
was most noticable because a further MVMCallCapture could then come to
refer to this same callsite, be marked, and the mark would cause a use
after free.

Making sure that it stays alive costs an extra pointer in MVMFrame,
which is a bit of a pity, but less of one than SIGSEGV.