Always copy callsite into an MVMCallCapture.
Rather than trying to reference it. That turns out to be reliably
fragile, and it's not a big thing to copy. The reward is that we get
to eliminate a bunch of logic, making for clearer code, and also lose
a flag that - after alignment considerations - will save 4 or 8 bytes
per call capture.