Code-gen and install planned specializations.
With this, spesh now starts to work again, with other threads able to
run the specializations generated by the worker thread. It still only
does a single collection/planning run, however, and all of the logged
data is not yet being incorporated into the facts yet. OSR will also
need to be put back in place.