Remove synchronous candidate production from OSR.
This re-works the OSR code in preparation for it obtaining specialized
code produced by the worker thread. The installation code remains as
it was, though it's easier now as we don't have to do a logging and
then a final installation. The polling for a result will come later;
this just removes OSR's calls to the previous specialization handling.