Enable collection of spesh log data over time.
This establishes a kind of quota mechanism so that we can fill up to
3 (by default) buffers of log data per thread. The thread sending will
decrement the quota when it sends a buffer. Once the spesh worker
has processed a buffer, it increments the quota again. This limits the
amount of logged data around at the same time, but allows us to keep
accumulating data and doing more specialization over time.