When concat needs normalization only normalize one section
Previously if the last grapheme of string a and the first grapheme
of string b determined that renormalization was needed to combine these
codepoints, we were forced to renormalize all of string a and all of
string b, together into one string.

When we concat normally, we create a string which contains the string
or strands of string a and the string or strands of string b together
in one string. With the new code we copy the strands or set string a
as the first strand of the new string, then if we have a renormalized
section to insert afterward, we adjust the last strand of the copied
strands from string a or the single string of string a to refer to a
substring which is all but the last grapheme.

Then we insert the renormalized section if it exists. Then we copy the
strands or set string b as the last strand of our new string. We then
adjust the first strand of the copied section of string b or adjust the
string b's single strand to be a substring.

In addition to this, we "drop" a copied strand or string a or string b
if the adjusted substring has no more graphemes in it. This currently
can only happen if string a or the last strand of string a is a single
grapheme or the first strand of string b or string b is a single grapheme.

In the future we may cover more cases, but for now we only cover the
case of the two codepoints combining and don't use this method unless
both last_a and first_b are non-synthetics and both have a canonical
combining class of 0. In the future this may become less restrictive.