Ensure spesh log has correct thread.
Even when it's created by a thread other than the one its for. Prior
to this fix, we filled all the spesh logs, sent them off to the spesh
worker thread, which then after specializing created a log with the
wrong thread object in it. The upshot was that we never created any
more logs after that one for the logging/sending thread.