concat: don't do re_nfg shortcut if either a or b are repeats
I previously had to disable this for synthetics, but further testing
shows that the issue exists for any repeat. This is caused by repeat's
essentially acting differently than other strands. Firstly they are
not guaranteed to be normalized once combined and secondly the
code does not account for them properly.