Add type checks for templates
Many, many, many errors I've made by inserting a VOID-yielding node in
a template where a REG was required or the other way arround. And that
then doesn't show up until the template is tiled and you get a
segmentation fault or a panic. So let's try to avoid that and add node
type checks, which will find such problems at (pre)compilation time.