concat: Fix conditional to check first instead of last strand of string b
This is much likely to cause an issue than repetitions in the last strand of
string a, but correct this so it checks the proper strand.

I was not able to trigger any incorrectness so was unable to add new tests
for this change like I did for when string a has repetitions.