Fix bad guard assumptions around OSR points.
Prior to this, it was possible that a condition checked before an OSR
point would always hold after it. This was not a good assumption, as
at the point OSR happens the current value in the interpreted code may
not have met the guard.

To fix this, first ensure that the osrpoint instruction starts a new
basic block (this was typically the case anyway due to its placement).
Then make this a successor of the entry point, which is what we do for
excieption handlers also. The invalidation of guards will then fall
naturally out of fact merging at PHI nodes.