JIT should include SC write barriers.
Otherwise, when we are compiling things that rely on repossession of
serialized objects, that will not happen, resulting in mis-compiles.
It would be nice in the future ot only conditionally include these
(and specialize on us not being in the context of compilation). For
now, this at least gets things correct.