NULL out registers on re-OSR.
If we previously ran the OSR'd code for this frame, then we don't need
to resize work/env. However, they were unused while the deopt code was
being run, and so may contain outdated pointers that will upset the
GC if it comes across them. Fix this by making sure that space is
cleared out.