Simplify our dealings with the thread list.
Just take a lock for the cases where we need to update or snapshot it.
This eliminates a low-benefit, and difficult to reason about,
lock-free approach in the GC orchestration, making for far simpler
code there. It also disentangles it a bit from the GC start flag
handling, which will be getting a refactor shortly to try to improve
its resource usage.