Avoid many array searches for SC code refs.
This showed up hot in profiling. It turns out that we weren't even
checking if the code ref had a cached SC index, and even if we had
been we'd still get a miss because we didn't set the SC on the code
object before adding it, meaning the index was never cached. Cuts the
number of times we need to take the loop to just 40% of the cases it
used to need to in CORE.setting compilation, which cuts about 5% off
stage MAST.