Use a bit field for named args used when possible.
When there only 64 named arguments, we can use a bit field instead
of allocating a byte array. This saves an allocation, a memset to
zero it, and a free, in pretty much every use of named arguments (we
have to handle the case where there are more than 64 of them, but it
is likely incredibly rare).