Improve NFA memory performance.
We only need 32-bit arrays, not 64-bit ones; we'll never have an
NFA with billions of states! Even with that, the `memset` of the
`done` array turned out to account for a ghastly number of CPU cyles.
Since in most cases we get just in the single digits of things set
to `done`, it's cheaper to use an array for that and do lookups in it.

There are probably some further wins to be had in this area, but in
the meantime this gets 5% of the CPU cycles off the compilation of
CORE.setting.