Zero owner of a stack frame also.
The flags being zero is a sufficient test for whether the frame is a
stack frame or not when we know it's certainly a frame. However, when
we have an unknown collectable, the fully zeroed flags would also be
a normal object. The temp roots mechanism thus uses a zero owner to
identify this situation. Ensure the owner is zeroed.