Implement a ring buffer so only different codepoints need processing
The ring buffers hold the exact number of codepoints which comprise the longest
sequence of codepoints which map to its own collation keys in the Unicode
Collation Algorithm. As of Unicode 9.0 this number was 3.

This also allows us to return based on codepoint based on the determination
we found using the ring buffer