Fix MVM_string_compare to support deterministic comparing of synthetics
Previously we compared naively by grapheme, and ended up comparing synthetic
codepoints with non-synthetics. This would cause synthetics to be sorted
incorrectly, in addition to it making comparing things non-deterministic;
if the synthetics were added in a different order, you would get a different
result with MVM_string_compare.

Now, we compare down the string the same way we did before, but if the first
non-matching grapheme is a synthetic, we iterate over the codepoints in the
synthetic, so it returns the result based on codepoint. If the two non-matching
graphemes contain the same initial codepoints but different in length, the
grapheme with less codepoints in it is determined as less than the other.