Improve frame heap allocation prediction.
Only bump the counter of heap promotions when we have a spesh
correlation ID, which means we will have bumped the entries; before
this could get out of sync and over-state the frame's preference.
Also increaes the thresholds some. This greatly brings down the
number of added GC runs that this change caused by decreasing the
number of false positives.