Change the way we log returns.
The previous way was fragile and could lead to us getting bogus depth
counts, which then confused the order that specialization was done in
(not to mention anyone reading the spesh log). It also would block any
effort to try and carry over the simulated stack between processing of
logs.