Store deopt one target directly in bytecode.
Rather than needing to use the offsets to resolve them each deopt.
This brings the interpreted approach in line with what the JIT does,
but further to that it means that we can have a sequence of guards
that will deopt to the same place, which will be used for upcoming
improvements to invocation specialization.