When inlining, replace receive instruction.
If two parameters are received into the same register (because, for
example, they are then boxed and only the target of the box matters),
then the arg passing code would be rewritten in such a way that the
second argument overwrote the value of the first. This resolves the
issue by flipping the rewriting to the receiving side, where this
issue cannot happen.